Set background-color only on elements with existing background colors.
Previously we set the background-color on just the body and made
other elements transparent. That change caused issues with
drop-down boxes and modals going transparent.

Now, we set the background-color on any elements with an existing
background color. We have to use JavaScript to mark those elements
for our CSS styles due to lack of an appropriate CSS selector.